1. Модульные тесты. [#159435]

+++ src/ru/job4j/converter/Converter.java
@@ -2,18 +2,18 @@ package ru.job4j.converter;
 
 public class Converter {
 
-    public static int rubleToEuro(int value) {
+    public static float rubleToEuro(float value) {
         return value / 70;
     }
 
-    public static int rubleToDollar(int value) {
+    public static float rubleToDollar(float value) {
         return value / 60;
     }
 
     public static void main(String[] args) {
-        int in = 140;
-        int expected = 2;
-        int out = Converter.rubleToEuro(in);
+        float in = 140;
+        float expected = 2;
+        float out = Converter.rubleToEuro(in);
         boolean passed = expected == out;
         System.out.println("140 rubles are 2. Test result : " + passed);
         in = 280;
